+++ genotype_api/api/endpoints/analyses.py
@@ -15,7 +15,8 @@ from genotype_api.database.crud.read import (
     get_analyses_with_skip_and_limit,
 )
 from genotype_api.database.crud.update import refresh_sample_status
-from genotype_api.database.models import Analysis, AnalysisRead, AnalysisReadWithGenotype, User
+from genotype_api.database.models import Analysis, User
+from genotype_api.dto.dto import AnalysisRead, AnalysisReadWithGenotype
 from genotype_api.database.session_handler import get_session
 from genotype_api.file_parsing.files import check_file
 from genotype_api.file_parsing.vcf import SequenceAnalysis
